Uses hash tables to detect duplicates

+++ parseh5.py
@@ -13,16 +13,31 @@ class VehicleLocation(tables.IsDescription):
     time      = tables.Float64Col()   #Time stamp in seconds since epoch time
     heading   = tables.UInt16Col()    #Heading in degrees
 
-def parse_mbta_xml(database, thefile):
+def parse_mbta_xml(database, thefile, presentData = None):
+    """
+    Parses MBTA XML data and adds it to a HDF5 database.
+
+    Inputs:
+    database: Handle to HDF5 file
+    thefile: Name of XML file to parse
+    presentData: A dictionary hash of present data (to save time on the check)
+                 If absent, will use database queries (much slower)
+    """
     try:
         tree = ET.parse(thefile)
         root = tree.getroot()
     except ET.ParseError: #Error parsing XML content of the file
+        print 'Could not find root of XML file', thefile
         return
 
     #Part 1. Get epoch time to nearest second
     #        MBTA reports in whole units of milliseconds
-    report_time = long(root.find('lastTime').attrib['time'][:-3])
+    timeData = root.find('lastTime')
+    if timeData is None: #Maybe XML returned an error of some sort
+        print 'XML file %s does not have time data' % thefile
+        return
+
+    report_time = long(timeData.attrib['time'][:-3])
     
     #Part 2. Parse vehicle location data.
     for thevehicle in root.iter('vehicle'):
@@ -41,32 +56,44 @@ def parse_mbta_xml(database, thefile):
         except KeyError:
             pass
         
-        #Check that this record wasn't already reported
-        queryString = '((vehicleID == "%(vehicleID)s") & (time == %(time)s))' % data 
-        try:
-            query = database.getWhereList(queryString)
-        except tables.exceptions.HDF5ExtError:
-            #gets thrown whenHDF5 file is open and being written to
-            print "Could not get file lock on HDF5 file. Abort."
-            import sys
-            sys.exit()
+        #Part 3. Make sure record is not a duplicate
+        if presentData is None:
+            #No hashes. Query database to check that this record wasn't already reported
+            queryString = '((vehicleID == "%(vehicleID)s") & (time == %(time)s))' % data 
+            try:
+                query = database.getWhereList(queryString)
+            except tables.exceptions.HDF5ExtError:
+                #gets thrown whenHDF5 file is open and being written to
+                print "Could not get file lock on HDF5 file. Abort."
+                import sys
+                sys.exit()
 
-        if len(query) == 0:
-            vehiclePosition = database.row
-            for key, value in data.items():
-                vehiclePosition[key] = value
-            vehiclePosition.append()
+            if len(query) == 0:
+                vehiclePosition = database.row
+                for key, value in data.items():
+                    vehiclePosition[key] = value
+                vehiclePosition.append()
+            else:
+                assert len(query) == 1, "OMG COLLISION"
         else:
-            assert len(query) == 1, "OMG COLLISION"
+            #Use hashes to check if record is already reported
+            if (data['vehicleID'], data['time']) not in presentData:
+                vehiclePosition = database.row
+                for key, value in data.items():
+                    vehiclePosition[key] = value
+                vehiclePosition.append()
+                presentData[data['vehicleID'], data['time']] = True
+
     database.flush()
+    print 'Parsed', thefile
+    return presentData
 
 
-
-def ParseAll():
+def ParseAll(theHDF5FileName = 'mbta_trajectories.h5', Cleanup = True):
     import glob, os
 
     compressionOptions = tables.Filters(complevel=9, complib='blosc')
-    f = tables.openFile('mbta_trajectories.h5', mode = 'a',
+    f = tables.openFile(theHDF5FileName, mode = 'a',
         filters = compressionOptions, title = 'Historical MBTA bus data')
    
     try:
@@ -78,11 +105,15 @@ def ParseAll():
         thetable.cols.time.createIndex()
         thetable.cols.vehicleID.createIndex()
 
+    #Hash current data
+    presentData = {}
+    for row in thetable:
+        presentData[row['vehicleID'], row['time']] = True
 
     for filename in sorted(glob.glob('*.xml')):
-        parse_mbta_xml(thetable, filename)
-        print 'Parsed', filename
-        os.unlink(filename)
+        presentData = parse_mbta_xml(thetable, filename, presentData)
+        if Cleanup:
+            os.unlink(filename)
 
     f.close()
